v.0.5.2: Reducing overall memory consumption while simulating.

+++ include/collections/sstream.h
@@ -5,6 +5,9 @@
 
 #define SSLEN 1024
 
+// C Headers
+#include <inttypes.h>
+
 typedef struct _sstream {
 
     int length;
@@ -15,8 +18,8 @@ typedef struct _sstream {
 extern void sstream_ctor(struct _sstream *this);
 
 extern void sstream_put(struct _sstream *this, const char *fmt, ...);
-extern void sstream_put04x(struct _sstream *this, const int decimal);
-extern void sstream_put02x(struct _sstream *this, const int decimal);
+extern void sstream_put04x(struct _sstream *this, const unsigned int decimal);
+extern void sstream_put02x(struct _sstream *this, const uint8_t decimal);
 
 extern void sstream_pad(struct _sstream *this, const int padding);
 extern char* sstream_alloc(const struct _sstream *this);
